Struct aws_sdk_s3control::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSS3ControlServiceV20180820.

This client allows ergonomic access to a AWSS3ControlServiceV20180820-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_access_point(&self) -> CreateAccessPoint<C, M, R>

Constructs a fluent builder for the CreateAccessPoint operation.

See CreateAccessPoint for more information about the operation and its arguments.

pub fn create_access_point_for_object_lambda(
&self
) -> CreateAccessPointForObjectLambda<C, M, R>

Constructs a fluent builder for the CreateAccessPointForObjectLambda operation.

See CreateAccessPointForObjectLambda for more information about the operation and its arguments.

pub fn create_bucket(&self) -> CreateBucket<C, M, R>

Constructs a fluent builder for the CreateBucket operation.

See CreateBucket for more information about the operation and its arguments.

pub fn create_job(&self) -> CreateJob<C, M, R>

Constructs a fluent builder for the CreateJob operation.

See CreateJob for more information about the operation and its arguments.

pub fn create_multi_region_access_point(
&self
) -> CreateMultiRegionAccessPoint<C, M, R>

Constructs a fluent builder for the CreateMultiRegionAccessPoint operation.

See CreateMultiRegionAccessPoint for more information about the operation and its arguments.

pub fn delete_access_point(&self) -> DeleteAccessPoint<C, M, R>

Constructs a fluent builder for the DeleteAccessPoint operation.

See DeleteAccessPoint for more information about the operation and its arguments.

pub fn delete_access_point_for_object_lambda(
&self
) -> DeleteAccessPointForObjectLambda<C, M, R>

Constructs a fluent builder for the DeleteAccessPointForObjectLambda operation.

See DeleteAccessPointForObjectLambda for more information about the operation and its arguments.

pub fn delete_access_point_policy(&self) -> DeleteAccessPointPolicy<C, M, R>

Constructs a fluent builder for the DeleteAccessPointPolicy operation.

See DeleteAccessPointPolicy for more information about the operation and its arguments.

pub fn delete_access_point_policy_for_object_lambda(
&self
) -> DeleteAccessPointPolicyForObjectLambda<C, M, R>

Constructs a fluent builder for the DeleteAccessPointPolicyForObjectLambda operation.

See DeleteAccessPointPolicyForObjectLambda for more information about the operation and its arguments.

pub fn delete_bucket(&self) -> DeleteBucket<C, M, R>

Constructs a fluent builder for the DeleteBucket operation.

See DeleteBucket for more information about the operation and its arguments.

pub fn delete_bucket_lifecycle_configuration(
&self
) -> DeleteBucketLifecycleConfiguration<C, M, R>

Constructs a fluent builder for the DeleteBucketLifecycleConfiguration operation.

See DeleteBucketLifecycleConfiguration for more information about the operation and its arguments.

pub fn delete_bucket_policy(&self) -> DeleteBucketPolicy<C, M, R>

Constructs a fluent builder for the DeleteBucketPolicy operation.

See DeleteBucketPolicy for more information about the operation and its arguments.

pub fn delete_bucket_tagging(&self) -> DeleteBucketTagging<C, M, R>

Constructs a fluent builder for the DeleteBucketTagging operation.

See DeleteBucketTagging for more information about the operation and its arguments.

pub fn delete_job_tagging(&self) -> DeleteJobTagging<C, M, R>

Constructs a fluent builder for the DeleteJobTagging operation.

See DeleteJobTagging for more information about the operation and its arguments.

pub fn delete_multi_region_access_point(
&self
) -> DeleteMultiRegionAccessPoint<C, M, R>

Constructs a fluent builder for the DeleteMultiRegionAccessPoint operation.

See DeleteMultiRegionAccessPoint for more information about the operation and its arguments.

pub fn delete_public_access_block(&self) -> DeletePublicAccessBlock<C, M, R>

Constructs a fluent builder for the DeletePublicAccessBlock operation.

See DeletePublicAccessBlock for more information about the operation and its arguments.

pub fn delete_storage_lens_configuration(
&self
) -> DeleteStorageLensConfiguration<C, M, R>

Constructs a fluent builder for the DeleteStorageLensConfiguration operation.

See DeleteStorageLensConfiguration for more information about the operation and its arguments.

pub fn delete_storage_lens_configuration_tagging(
&self
) -> DeleteStorageLensConfigurationTagging<C, M, R>

Constructs a fluent builder for the DeleteStorageLensConfigurationTagging operation.

See DeleteStorageLensConfigurationTagging for more information about the operation and its arguments.

pub fn describe_job(&self) -> DescribeJob<C, M, R>

Constructs a fluent builder for the DescribeJob operation.

See DescribeJob for more information about the operation and its arguments.

pub fn describe_multi_region_access_point_operation(
&self
) -> DescribeMultiRegionAccessPointOperation<C, M, R>

Constructs a fluent builder for the DescribeMultiRegionAccessPointOperation operation.

See DescribeMultiRegionAccessPointOperation for more information about the operation and its arguments.

pub fn get_access_point(&self) -> GetAccessPoint<C, M, R>

Constructs a fluent builder for the GetAccessPoint operation.

See GetAccessPoint for more information about the operation and its arguments.

pub fn get_access_point_configuration_for_object_lambda(
&self
) -> GetAccessPointConfigurationForObjectLambda<C, M, R>

Constructs a fluent builder for the GetAccessPointConfigurationForObjectLambda operation.

See GetAccessPointConfigurationForObjectLambda for more information about the operation and its arguments.

pub fn get_access_point_for_object_lambda(
&self
) -> GetAccessPointForObjectLambda<C, M, R>

Constructs a fluent builder for the GetAccessPointForObjectLambda operation.

See GetAccessPointForObjectLambda for more information about the operation and its arguments.

pub fn get_access_point_policy(&self) -> GetAccessPointPolicy<C, M, R>

Constructs a fluent builder for the GetAccessPointPolicy operation.

See GetAccessPointPolicy for more information about the operation and its arguments.

pub fn get_access_point_policy_for_object_lambda(
&self
) -> GetAccessPointPolicyForObjectLambda<C, M, R>

Constructs a fluent builder for the GetAccessPointPolicyForObjectLambda operation.

See GetAccessPointPolicyForObjectLambda for more information about the operation and its arguments.

pub fn get_access_point_policy_status(
&self
) -> GetAccessPointPolicyStatus<C, M, R>

Constructs a fluent builder for the GetAccessPointPolicyStatus operation.

See GetAccessPointPolicyStatus for more information about the operation and its arguments.

pub fn get_access_point_policy_status_for_object_lambda(
&self
) -> GetAccessPointPolicyStatusForObjectLambda<C, M, R>

Constructs a fluent builder for the GetAccessPointPolicyStatusForObjectLambda operation.

See GetAccessPointPolicyStatusForObjectLambda for more information about the operation and its arguments.

pub fn get_bucket(&self) -> GetBucket<C, M, R>

Constructs a fluent builder for the GetBucket operation.

See GetBucket for more information about the operation and its arguments.

pub fn get_bucket_lifecycle_configuration(
&self
) -> GetBucketLifecycleConfiguration<C, M, R>

Constructs a fluent builder for the GetBucketLifecycleConfiguration operation.

See GetBucketLifecycleConfiguration for more information about the operation and its arguments.

pub fn get_bucket_policy(&self) -> GetBucketPolicy<C, M, R>

Constructs a fluent builder for the GetBucketPolicy operation.

See GetBucketPolicy for more information about the operation and its arguments.

pub fn get_bucket_tagging(&self) -> GetBucketTagging<C, M, R>

Constructs a fluent builder for the GetBucketTagging operation.

See GetBucketTagging for more information about the operation and its arguments.

pub fn get_job_tagging(&self) -> GetJobTagging<C, M, R>

Constructs a fluent builder for the GetJobTagging operation.

See GetJobTagging for more information about the operation and its arguments.

pub fn get_multi_region_access_point(
&self
) -> GetMultiRegionAccessPoint<C, M, R>

Constructs a fluent builder for the GetMultiRegionAccessPoint operation.

See GetMultiRegionAccessPoint for more information about the operation and its arguments.

pub fn get_multi_region_access_point_policy(
&self
) -> GetMultiRegionAccessPointPolicy<C, M, R>

Constructs a fluent builder for the GetMultiRegionAccessPointPolicy operation.

See GetMultiRegionAccessPointPolicy for more information about the operation and its arguments.

pub fn get_multi_region_access_point_policy_status(
&self
) -> GetMultiRegionAccessPointPolicyStatus<C, M, R>

Constructs a fluent builder for the GetMultiRegionAccessPointPolicyStatus operation.

See GetMultiRegionAccessPointPolicyStatus for more information about the operation and its arguments.

pub fn get_public_access_block(&self) -> GetPublicAccessBlock<C, M, R>

Constructs a fluent builder for the GetPublicAccessBlock operation.

See GetPublicAccessBlock for more information about the operation and its arguments.

pub fn get_storage_lens_configuration(
&self
) -> GetStorageLensConfiguration<C, M, R>

Constructs a fluent builder for the GetStorageLensConfiguration operation.

See GetStorageLensConfiguration for more information about the operation and its arguments.

pub fn get_storage_lens_configuration_tagging(
&self
) -> GetStorageLensConfigurationTagging<C, M, R>

Constructs a fluent builder for the GetStorageLensConfigurationTagging operation.

See GetStorageLensConfigurationTagging for more information about the operation and its arguments.

pub fn list_access_points(&self) -> ListAccessPoints<C, M, R>

Constructs a fluent builder for the ListAccessPoints operation.

See ListAccessPoints for more information about the operation and its arguments.

pub fn list_access_points_for_object_lambda(
&self
) -> ListAccessPointsForObjectLambda<C, M, R>

Constructs a fluent builder for the ListAccessPointsForObjectLambda operation.

See ListAccessPointsForObjectLambda for more information about the operation and its arguments.

pub fn list_jobs(&self) -> ListJobs<C, M, R>

Constructs a fluent builder for the ListJobs operation.

See ListJobs for more information about the operation and its arguments.

pub fn list_multi_region_access_points(
&self
) -> ListMultiRegionAccessPoints<C, M, R>

Constructs a fluent builder for the ListMultiRegionAccessPoints operation.

See ListMultiRegionAccessPoints for more information about the operation and its arguments.

pub fn list_regional_buckets(&self) -> ListRegionalBuckets<C, M, R>

Constructs a fluent builder for the ListRegionalBuckets operation.

See ListRegionalBuckets for more information about the operation and its arguments.

pub fn list_storage_lens_configurations(
&self
) -> ListStorageLensConfigurations<C, M, R>

Constructs a fluent builder for the ListStorageLensConfigurations operation.

See ListStorageLensConfigurations for more information about the operation and its arguments.

pub fn put_access_point_configuration_for_object_lambda(
&self
) -> PutAccessPointConfigurationForObjectLambda<C, M, R>

Constructs a fluent builder for the PutAccessPointConfigurationForObjectLambda operation.

See PutAccessPointConfigurationForObjectLambda for more information about the operation and its arguments.

pub fn put_access_point_policy(&self) -> PutAccessPointPolicy<C, M, R>

Constructs a fluent builder for the PutAccessPointPolicy operation.

See PutAccessPointPolicy for more information about the operation and its arguments.

pub fn put_access_point_policy_for_object_lambda(
&self
) -> PutAccessPointPolicyForObjectLambda<C, M, R>

Constructs a fluent builder for the PutAccessPointPolicyForObjectLambda operation.

See PutAccessPointPolicyForObjectLambda for more information about the operation and its arguments.

pub fn put_bucket_lifecycle_configuration(
&self
) -> PutBucketLifecycleConfiguration<C, M, R>

Constructs a fluent builder for the PutBucketLifecycleConfiguration operation.

See PutBucketLifecycleConfiguration for more information about the operation and its arguments.

pub fn put_bucket_policy(&self) -> PutBucketPolicy<C, M, R>

Constructs a fluent builder for the PutBucketPolicy operation.

See PutBucketPolicy for more information about the operation and its arguments.

pub fn put_bucket_tagging(&self) -> PutBucketTagging<C, M, R>

Constructs a fluent builder for the PutBucketTagging operation.

See PutBucketTagging for more information about the operation and its arguments.

pub fn put_job_tagging(&self) -> PutJobTagging<C, M, R>

Constructs a fluent builder for the PutJobTagging operation.

See PutJobTagging for more information about the operation and its arguments.

pub fn put_multi_region_access_point_policy(
&self
) -> PutMultiRegionAccessPointPolicy<C, M, R>

Constructs a fluent builder for the PutMultiRegionAccessPointPolicy operation.

See PutMultiRegionAccessPointPolicy for more information about the operation and its arguments.

pub fn put_public_access_block(&self) -> PutPublicAccessBlock<C, M, R>

Constructs a fluent builder for the PutPublicAccessBlock operation.

See PutPublicAccessBlock for more information about the operation and its arguments.

pub fn put_storage_lens_configuration(
&self
) -> PutStorageLensConfiguration<C, M, R>

Constructs a fluent builder for the PutStorageLensConfiguration operation.

See PutStorageLensConfiguration for more information about the operation and its arguments.

pub fn put_storage_lens_configuration_tagging(
&self
) -> PutStorageLensConfigurationTagging<C, M, R>

Constructs a fluent builder for the PutStorageLensConfigurationTagging operation.

See PutStorageLensConfigurationTagging for more information about the operation and its arguments.

pub fn update_job_priority(&self) -> UpdateJobPriority<C, M, R>

Constructs a fluent builder for the UpdateJobPriority operation.

See UpdateJobPriority for more information about the operation and its arguments.

pub fn update_job_status(&self) -> UpdateJobStatus<C, M, R>

Constructs a fluent builder for the UpdateJobStatus operation.

See UpdateJobStatus for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.