Module aws_sdk_s3control::error

Modules

bad_request_exception

See BadRequestException

bucket_already_exists

See BucketAlreadyExists

bucket_already_owned_by_you

See BucketAlreadyOwnedByYou

idempotency_exception

See IdempotencyException

internal_service_exception

See InternalServiceException

invalid_next_token_exception

See InvalidNextTokenException

invalid_request_exception

See InvalidRequestException

job_status_exception

See JobStatusException

no_such_public_access_block_configuration

See NoSuchPublicAccessBlockConfiguration

not_found_exception

See NotFoundException

too_many_requests_exception

See TooManyRequestsException

too_many_tags_exception

See TooManyTagsException

Structs

BadRequestException

BucketAlreadyExists

The requested Outposts bucket name is not available. The bucket namespace is shared by

BucketAlreadyOwnedByYou

The Outposts bucket you tried to create already exists, and you own it.

CreateAccessPointError

CreateAccessPointForObjectLambdaError

CreateBucketError

CreateJobError

CreateMultiRegionAccessPointError

DeleteAccessPointError

DeleteAccessPointForObjectLambdaError

DeleteAccessPointPolicyError

DeleteAccessPointPolicyForObjectLambdaError

DeleteBucketError

DeleteBucketLifecycleConfigurationError

DeleteBucketPolicyError

DeleteBucketTaggingError

DeleteJobTaggingError

DeleteMultiRegionAccessPointError

DeletePublicAccessBlockError

DeleteStorageLensConfigurationError

DeleteStorageLensConfigurationTaggingError

DescribeJobError

DescribeMultiRegionAccessPointOperationError

GetAccessPointConfigurationForObjectLambdaError

GetAccessPointError

GetAccessPointForObjectLambdaError

GetAccessPointPolicyError

GetAccessPointPolicyForObjectLambdaError

GetAccessPointPolicyStatusError

GetAccessPointPolicyStatusForObjectLambdaError

GetBucketError

GetBucketLifecycleConfigurationError

GetBucketPolicyError

GetBucketTaggingError

GetJobTaggingError

GetMultiRegionAccessPointError

GetMultiRegionAccessPointPolicyError

GetMultiRegionAccessPointPolicyStatusError

GetPublicAccessBlockError

GetStorageLensConfigurationError

GetStorageLensConfigurationTaggingError

IdempotencyException

InternalServiceException

InvalidNextTokenException

InvalidRequestException

JobStatusException

ListAccessPointsError

ListAccessPointsForObjectLambdaError

ListJobsError

ListMultiRegionAccessPointsError

ListRegionalBucketsError

ListStorageLensConfigurationsError

NoSuchPublicAccessBlockConfiguration

Amazon S3 throws this exception if you make a GetPublicAccessBlock request

NotFoundException

PutAccessPointConfigurationForObjectLambdaError

PutAccessPointPolicyError

PutAccessPointPolicyForObjectLambdaError

PutBucketLifecycleConfigurationError

PutBucketPolicyError

PutBucketTaggingError

PutJobTaggingError

PutMultiRegionAccessPointPolicyError

PutPublicAccessBlockError

PutStorageLensConfigurationError

PutStorageLensConfigurationTaggingError

TooManyRequestsException

TooManyTagsException

Amazon S3 throws this exception if you have too many tags in your tag set.

UpdateJobPriorityError

UpdateJobStatusError

Enums

CreateAccessPointErrorKind

CreateAccessPointForObjectLambdaErrorKind

CreateBucketErrorKind

CreateJobErrorKind

CreateMultiRegionAccessPointErrorKind

DeleteAccessPointErrorKind

DeleteAccessPointForObjectLambdaErrorKind

DeleteAccessPointPolicyErrorKind

DeleteAccessPointPolicyForObjectLambdaErrorKind

DeleteBucketErrorKind

DeleteBucketLifecycleConfigurationErrorKind

DeleteBucketPolicyErrorKind

DeleteBucketTaggingErrorKind

DeleteJobTaggingErrorKind

DeleteMultiRegionAccessPointErrorKind

DeletePublicAccessBlockErrorKind

DeleteStorageLensConfigurationErrorKind

DeleteStorageLensConfigurationTaggingErrorKind

DescribeJobErrorKind

DescribeMultiRegionAccessPointOperationErrorKind

GetAccessPointConfigurationForObjectLambdaErrorKind

GetAccessPointErrorKind

GetAccessPointForObjectLambdaErrorKind

GetAccessPointPolicyErrorKind

GetAccessPointPolicyForObjectLambdaErrorKind

GetAccessPointPolicyStatusErrorKind

GetAccessPointPolicyStatusForObjectLambdaErrorKind

GetBucketErrorKind

GetBucketLifecycleConfigurationErrorKind

GetBucketPolicyErrorKind

GetBucketTaggingErrorKind

GetJobTaggingErrorKind

GetMultiRegionAccessPointErrorKind

GetMultiRegionAccessPointPolicyErrorKind

GetMultiRegionAccessPointPolicyStatusErrorKind

GetPublicAccessBlockErrorKind

GetStorageLensConfigurationErrorKind

GetStorageLensConfigurationTaggingErrorKind

ListAccessPointsErrorKind

ListAccessPointsForObjectLambdaErrorKind

ListJobsErrorKind

ListMultiRegionAccessPointsErrorKind

ListRegionalBucketsErrorKind

ListStorageLensConfigurationsErrorKind

PutAccessPointConfigurationForObjectLambdaErrorKind

PutAccessPointPolicyErrorKind

PutAccessPointPolicyForObjectLambdaErrorKind

PutBucketLifecycleConfigurationErrorKind

PutBucketPolicyErrorKind

PutBucketTaggingErrorKind

PutJobTaggingErrorKind

PutMultiRegionAccessPointPolicyErrorKind

PutPublicAccessBlockErrorKind

PutStorageLensConfigurationErrorKind

PutStorageLensConfigurationTaggingErrorKind

UpdateJobPriorityErrorKind

UpdateJobStatusErrorKind