Module aws_sdk_s3control::model

Modules

abort_incomplete_multipart_upload

See AbortIncompleteMultipartUpload

access_point

See AccessPoint

account_level

See AccountLevel

activity_metrics

See ActivityMetrics

async_error_details

See AsyncErrorDetails

async_operation

See AsyncOperation

async_request_parameters

See AsyncRequestParameters

async_response_details

See AsyncResponseDetails

aws_lambda_transformation

See AwsLambdaTransformation

bucket_level

See BucketLevel

create_bucket_configuration

See CreateBucketConfiguration

create_multi_region_access_point_input

See CreateMultiRegionAccessPointInput

delete_multi_region_access_point_input

See DeleteMultiRegionAccessPointInput

established_multi_region_access_point_policy

See EstablishedMultiRegionAccessPointPolicy

exclude

See Exclude

include

See Include

job_descriptor

See JobDescriptor

job_failure

See JobFailure

job_list_descriptor

See JobListDescriptor

job_manifest

See JobManifest

job_manifest_location

See JobManifestLocation

job_manifest_spec

See JobManifestSpec

job_operation

See JobOperation

job_progress_summary

See JobProgressSummary

job_report

See JobReport

lambda_invoke_operation

See LambdaInvokeOperation

lifecycle_configuration

See LifecycleConfiguration

lifecycle_expiration

See LifecycleExpiration

lifecycle_rule

See LifecycleRule

lifecycle_rule_and_operator

See LifecycleRuleAndOperator

lifecycle_rule_filter

See LifecycleRuleFilter

list_storage_lens_configuration_entry

See ListStorageLensConfigurationEntry

multi_region_access_point_policy_document

See MultiRegionAccessPointPolicyDocument

multi_region_access_point_regional_response

See MultiRegionAccessPointRegionalResponse

multi_region_access_point_report

See MultiRegionAccessPointReport

multi_region_access_points_async_response

See MultiRegionAccessPointsAsyncResponse

noncurrent_version_expiration

See NoncurrentVersionExpiration

noncurrent_version_transition

See NoncurrentVersionTransition

object_lambda_access_point

See ObjectLambdaAccessPoint

object_lambda_configuration

See ObjectLambdaConfiguration

object_lambda_transformation_configuration

See ObjectLambdaTransformationConfiguration

policy_status

See PolicyStatus

prefix_level

See PrefixLevel

prefix_level_storage_metrics

See PrefixLevelStorageMetrics

proposed_multi_region_access_point_policy

See ProposedMultiRegionAccessPointPolicy

public_access_block_configuration

See PublicAccessBlockConfiguration

put_multi_region_access_point_policy_input

See PutMultiRegionAccessPointPolicyInput

region

See Region

region_report

See RegionReport

regional_bucket

See RegionalBucket

s3_access_control_list

See S3AccessControlList

s3_access_control_policy

See S3AccessControlPolicy

s3_bucket_destination

See S3BucketDestination

s3_copy_object_operation

See S3CopyObjectOperation

s3_delete_object_tagging_operation

See S3DeleteObjectTaggingOperation

s3_grant

See S3Grant

s3_grantee

See S3Grantee

s3_initiate_restore_object_operation

See S3InitiateRestoreObjectOperation

s3_object_lock_legal_hold

See S3ObjectLockLegalHold

s3_object_metadata

See S3ObjectMetadata

s3_object_owner

See S3ObjectOwner

s3_retention

See S3Retention

s3_set_object_acl_operation

See S3SetObjectAclOperation

s3_set_object_legal_hold_operation

See S3SetObjectLegalHoldOperation

s3_set_object_retention_operation

See S3SetObjectRetentionOperation

s3_set_object_tagging_operation

See S3SetObjectTaggingOperation

s3_tag

See S3Tag

selection_criteria

See SelectionCriteria

ssekms

See Ssekms

sses3

See Sses3

storage_lens_aws_org

See StorageLensAwsOrg

storage_lens_configuration

See StorageLensConfiguration

storage_lens_data_export

See StorageLensDataExport

storage_lens_data_export_encryption

See StorageLensDataExportEncryption

storage_lens_tag

See StorageLensTag

tagging

See Tagging

transition

See Transition

vpc_configuration

See VpcConfiguration

Structs

AbortIncompleteMultipartUpload

The container for abort incomplete multipart upload

AccessPoint

An access point used to access a bucket.

AccountLevel

A container for the account level Amazon S3 Storage Lens configuration.

ActivityMetrics

A container for the activity metrics.

AsyncErrorDetails

Error details for the failed asynchronous operation.

AsyncOperation

A container for the information about an asynchronous operation.

AsyncRequestParameters

A container for the request parameters associated with an asynchronous request.

AsyncResponseDetails

A container for the response details that are returned when querying about an

AwsLambdaTransformation

Lambda function used to transform objects through an Object Lambda Access Point.

BucketLevel

A container for the bucket-level configuration.

CreateBucketConfiguration

The container for the bucket configuration.

CreateMultiRegionAccessPointInput

A container for the information associated with a

DeleteMultiRegionAccessPointInput

A container for the information associated with a

EstablishedMultiRegionAccessPointPolicy

The last established access control policy for a Multi-Region Access Point.

Exclude

A container for what Amazon S3 Storage Lens will exclude.

Include

A container for what Amazon S3 Storage Lens configuration includes.

JobDescriptor

A container element for the job configuration and status information returned by a Describe Job request.

JobFailure

If this job failed, this element indicates why the job failed.

JobListDescriptor

Contains the configuration and status information for a single job retrieved as part of a job list.

JobManifest

Contains the configuration information for a job's manifest.

JobManifestLocation

Contains the information required to locate a manifest object.

JobManifestSpec

Describes the format of a manifest. If the manifest is in CSV format, also describes the columns contained within the manifest.

JobOperation

The operation that you want this job to perform on every object listed in the manifest.

JobProgressSummary

Describes the total number of tasks that the specified job has started, the number of tasks that succeeded, and the number of tasks that failed.

JobReport

Contains the configuration parameters for a job-completion report.

LambdaInvokeOperation

Contains the configuration parameters for a Lambda Invoke operation.

LifecycleConfiguration

The container for the Outposts bucket lifecycle configuration.

LifecycleExpiration

The container of the Outposts bucket lifecycle expiration.

LifecycleRule

The container for the Outposts bucket lifecycle rule.

LifecycleRuleAndOperator

The container for the Outposts bucket lifecycle rule and operator.

LifecycleRuleFilter

The container for the filter of the lifecycle rule.

ListStorageLensConfigurationEntry

Part of ListStorageLensConfigurationResult. Each entry includes the

MultiRegionAccessPointPolicyDocument

The Multi-Region Access Point access control policy.

MultiRegionAccessPointRegionalResponse

Status information for a single Multi-Region Access Point Region.

MultiRegionAccessPointReport

A collection of statuses for a Multi-Region Access Point in the various Regions it supports.

MultiRegionAccessPointsAsyncResponse

The Multi-Region Access Point details that are returned when querying about an asynchronous request.

NoncurrentVersionExpiration

The container of the noncurrent version expiration.

NoncurrentVersionTransition

The container for the noncurrent version transition.

ObjectLambdaAccessPoint

An access point with an attached Lambda function used to access transformed data from an Amazon S3 bucket.

ObjectLambdaConfiguration

A configuration used when creating an Object Lambda Access Point.

ObjectLambdaTransformationConfiguration

A configuration used when creating an Object Lambda Access Point transformation.

PolicyStatus

Indicates whether this access point policy is public. For more information about how Amazon S3 evaluates policies to determine whether they are public, see

PrefixLevel

A container for the prefix-level configuration.

PrefixLevelStorageMetrics

A container for the prefix-level storage metrics for S3 Storage Lens.

ProposedMultiRegionAccessPointPolicy

The proposed access control policy for the Multi-Region Access Point.

PublicAccessBlockConfiguration

The PublicAccessBlock configuration that you want to apply to this Amazon S3 account.

PutMultiRegionAccessPointPolicyInput

A container for the information associated with a

Region

A Region that supports a Multi-Region Access Point as well as the associated bucket for the Region.

RegionReport

A combination of a bucket and Region that's part of a Multi-Region Access Point.

RegionalBucket

The container for the regional bucket.

S3AccessControlList

S3AccessControlPolicy

S3BucketDestination

A container for the bucket where the Amazon S3 Storage Lens metrics export files are located.

S3CopyObjectOperation

Contains the configuration parameters for a PUT Copy object operation. S3 Batch Operations passes every object to the underlying PUT Copy object

S3DeleteObjectTaggingOperation

Contains no configuration parameters because the DELETE Object tagging API only accepts the bucket name and key name as parameters, which are defined in the job's manifest.

S3Grant

S3Grantee

S3InitiateRestoreObjectOperation

Contains the configuration parameters for an S3 Initiate Restore Object job.

S3ObjectLockLegalHold

Whether S3 Object Lock legal hold will be applied to objects in an S3 Batch Operations

S3ObjectMetadata

S3ObjectOwner

S3Retention

Contains the S3 Object Lock retention mode to be applied to all objects in the

S3SetObjectAclOperation

Contains the configuration parameters for a Set Object ACL operation. S3 Batch Operations passes every object to the underlying PUT Object acl API.

S3SetObjectLegalHoldOperation

Contains the configuration for an S3 Object Lock legal hold operation that an

S3SetObjectRetentionOperation

Contains the configuration parameters for the Object Lock retention action for an

S3SetObjectTaggingOperation

Contains the configuration parameters for a Set Object Tagging operation. S3 Batch Operations passes every object to the underlying PUT Object tagging API.

S3Tag

SelectionCriteria

Ssekms

Sses3

StorageLensAwsOrg

The Amazon Web Services organization for your S3 Storage Lens.

StorageLensConfiguration

A container for the Amazon S3 Storage Lens configuration.

StorageLensDataExport

A container to specify the properties of your S3 Storage Lens metrics export, including the

StorageLensDataExportEncryption

A container for the encryption of the S3 Storage Lens metrics exports.

StorageLensTag

Tagging

Transition

Specifies when an object transitions to a specified storage class. For more information

VpcConfiguration

The virtual private cloud (VPC) configuration for an access point.

Enums

AsyncOperationName

BucketCannedAcl

BucketLocationConstraint

ExpirationStatus

Format

JobManifestFieldName

JobManifestFormat

JobReportFormat

JobReportScope

JobStatus

MultiRegionAccessPointStatus

NetworkOrigin

ObjectLambdaAllowedFeature

ObjectLambdaContentTransformation

ObjectLambdaTransformationConfigurationAction

OperationName

OutputSchemaVersion

RequestedJobStatus

S3CannedAccessControlList

S3GlacierJobTier

S3GranteeTypeIdentifier

S3MetadataDirective

S3ObjectLockLegalHoldStatus

S3ObjectLockMode

S3ObjectLockRetentionMode

S3Permission

S3SseAlgorithm

S3StorageClass

TransitionStorageClass