Module aws_sdk_s3control::operation

Structs

CreateAccessPoint

Creates an access point and associates it with the specified bucket. For more information, see

CreateAccessPointForObjectLambda

Creates an Object Lambda Access Point. For more information, see

CreateBucket

CreateJob

You can use S3 Batch Operations to perform large-scale batch actions on Amazon S3 objects. Batch Operations can run a

CreateMultiRegionAccessPoint

Creates a Multi-Region Access Point and associates it with the specified buckets. For more information about creating Multi-Region Access Points, see

DeleteAccessPoint

Deletes the specified access point.

DeleteAccessPointForObjectLambda

Deletes the specified Object Lambda Access Point.

DeleteAccessPointPolicy

Deletes the access point policy for the specified access point.

DeleteAccessPointPolicyForObjectLambda

Removes the resource policy for an Object Lambda Access Point.

DeleteBucket

DeleteBucketLifecycleConfiguration

DeleteBucketPolicy

DeleteBucketTagging

DeleteJobTagging

Removes the entire tag set from the specified S3 Batch Operations job. To use this operation,

DeleteMultiRegionAccessPoint

Deletes a Multi-Region Access Point. This action does not delete the buckets associated with the Multi-Region Access Point,

DeletePublicAccessBlock

Removes the PublicAccessBlock configuration for an Amazon Web Services account. For more

DeleteStorageLensConfiguration

Deletes the Amazon S3 Storage Lens configuration. For more information about S3 Storage Lens, see

DeleteStorageLensConfigurationTagging

Deletes the Amazon S3 Storage Lens configuration tags. For more information about S3 Storage Lens, see

DescribeJob

Retrieves the configuration parameters and status for a Batch Operations job. For more

DescribeMultiRegionAccessPointOperation

Retrieves the status of an asynchronous request to manage a Multi-Region Access Point. For more information

GetAccessPoint

Returns configuration information about the specified access point.

GetAccessPointConfigurationForObjectLambda

Returns configuration for an Object Lambda Access Point.

GetAccessPointForObjectLambda

Returns configuration information about the specified Object Lambda Access Point

GetAccessPointPolicy

Returns the access point policy associated with the specified access point.

GetAccessPointPolicyForObjectLambda

Returns the resource policy for an Object Lambda Access Point.

GetAccessPointPolicyStatus

Indicates whether the specified access point currently has a policy that allows public access. For more information about public access through access points, see Managing Data Access with Amazon S3 access points in the Amazon S3 User Guide.

GetAccessPointPolicyStatusForObjectLambda

Returns the status of the resource policy associated with an Object Lambda Access Point.

GetBucket

Gets an Amazon S3 on Outposts bucket. For more information, see

GetBucketLifecycleConfiguration

GetBucketPolicy

GetBucketTagging

GetJobTagging

Returns the tags on an S3 Batch Operations job. To use this operation, you must have

GetMultiRegionAccessPoint

Returns configuration information about the specified Multi-Region Access Point.

GetMultiRegionAccessPointPolicy

Returns the access control policy of the specified Multi-Region Access Point.

GetMultiRegionAccessPointPolicyStatus

Indicates whether the specified Multi-Region Access Point has an access control policy that allows public

GetPublicAccessBlock

Retrieves the PublicAccessBlock configuration for an Amazon Web Services account. For more

GetStorageLensConfiguration

Gets the Amazon S3 Storage Lens configuration. For more information, see

GetStorageLensConfigurationTagging

Gets the tags of Amazon S3 Storage Lens configuration. For more information about S3 Storage Lens, see

ListAccessPoints

Returns a list of the access points currently associated with the specified bucket. You

ListAccessPointsForObjectLambda

Returns a list of the access points associated with the Object Lambda Access Point. You

ListJobs

Lists current S3 Batch Operations jobs and jobs that have ended within the last 30 days for

ListMultiRegionAccessPoints

Returns a list of the Multi-Region Access Points currently associated with the specified Amazon Web Services account. Each

ListRegionalBuckets

Returns a list of all Outposts buckets in an Outpost that are owned by the authenticated

ListStorageLensConfigurations

Gets a list of Amazon S3 Storage Lens configurations. For more information about S3 Storage Lens, see

PutAccessPointConfigurationForObjectLambda

Replaces configuration for an Object Lambda Access Point.

PutAccessPointPolicy

Associates an access policy with the specified access point. Each access point can have only one policy, so a request made to this API

PutAccessPointPolicyForObjectLambda

Creates or replaces resource policy for an Object Lambda Access Point. For an example policy, see

PutBucketLifecycleConfiguration

PutBucketPolicy

PutBucketTagging

PutJobTagging

Sets the supplied tag-set on an S3 Batch Operations job.

PutMultiRegionAccessPointPolicy

Associates an access control policy with the specified Multi-Region Access Point. Each Multi-Region Access Point can have only

PutPublicAccessBlock

Creates or modifies the PublicAccessBlock configuration for an Amazon Web Services account. For

PutStorageLensConfiguration

Puts an Amazon S3 Storage Lens configuration. For more information about S3 Storage Lens, see Working with

PutStorageLensConfigurationTagging

Put or replace tags on an existing Amazon S3 Storage Lens configuration. For more information about S3 Storage Lens, see

UpdateJobPriority

Updates an existing S3 Batch Operations job's priority. For more information, see S3 Batch Operations in the

UpdateJobStatus

Updates the status for the specified job. Use this action to confirm that you want to